Split knit remote.rs into client/push/clone/pull modules
Break the 1,411-line src/commands/remote.rs into client (HTTP transport over
curl, remote/token/config resolution, project-export fetch, bundle
localization, and shared feature-branch checkout helpers), push (remote config
CRUD plus project/bundle push and sync-on-push), clone (import a project export
into a fresh workspace), and pull (pull/fetch recorded bundle state and remote
bundle cleanup). remote/mod.rs holds the shared response DTOs and re-exports the
public API. No behavior change; tests and clippy unchanged.


Knit-Group: kg_20260531_43cd8e
Knit-Bundle: split-remote-and-gloss-cli
